Features/add alarms

+++ ApolloBVM/src/nscdrrn001pdunv.cpp
@@ -6,10 +6,10 @@ Nscdrrn001pdunv::Nscdrrn001pdunv(uint8_t pin, float pressure_conversion_multipli
     _pressure_conversion_constant(pressure_conversion_constant)
 {}
 
-float Nscdrrn001pdunv::read(){
+int Nscdrrn001pdunv::read(){
     return convertToPressure(analogRead(_pin));
 }
 
-float Nscdrrn001pdunv::convertToPressure(int input){
-    return (_pressure_conversion_multiplier * input) + _pressure_conversion_constant;
+int Nscdrrn001pdunv::convertToPressure(int input){
+    return (int) ( (_pressure_conversion_multiplier * input) + _pressure_conversion_constant );
 }
